faces/pyproject.toml
[project]
name = "faces"
version = "0.1.0"
description = "A face detection application that recognizes the faces of the team member"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "pycocotools>=2.0.0",
    "gdown>=5.0.0",
    "Pillow>=10.0.0",
    "filelock>=3.13.0",
    "beautifulsoup4>=4.12.0",
    "datasets>=2.14.0",
    "huggingface_hub>=0.19.0",
    "keras>=3.0.0",
    "tensorflow>=2.15.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "torch>=2.11.0",
]
